+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.2.3
+
+- print configuration hierarchy added, printing secrets disabled by default
+- replace parsing NODE_ENV from env only, added env configuration hierarchy with options.loadFilesFromEnv
+
 ## 1.2.1
 
 - cleanup eslint, fix support for markdown